pyproject.toml
[tool.poetry]
name = "ghostty-discord-bot"
version = "0.1.0"
description = ""
authors = ["Mitchell Hashimoto <m@mitchellh.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"

[tool.ruff.lint]
extend-select = ["I", "TID252", "SIM"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
pygithub = "^2.3.0"
sentry-sdk = "^2.3.1"
